requirements.txt
streamlit
chromadb
google-generativeai
requests
pdfplumber
python-docx
pydocx
pandas
openpyxl
pysqlite3-binary